Trekk ut felles logikk i egen VurderLivshendelseService
